+++ apps/finance/views.py
@@ -13,26 +13,21 @@ from apps.finance.serializers import (
     JournalSerializer,
     LedgerSerializer,
 )
+from apps.authentication.permissions import BusinessAccessMixin
 
 
-class BusinessOwnedMixin:
+class BusinessOwnedMixin(BusinessAccessMixin):
     permission_classes = [IsAuthenticated]
-
-    def get_business(self):
-        return get_object_or_404(
-            Business.objects.filter(owner=self.request.user),
-            pk=self.kwargs["business_id"],
-        )
 
 
 class AccountListView(BusinessOwnedMixin, APIView):
     def get(self, request, business_id):
-        self.get_business()
+        self.require_business_permission("finance", "view")
         qs = Account.objects.filter(business_id=business_id)
         return Response(AccountSerializer(qs, many=True).data)
 
     def post(self, request, business_id):
-        business = self.get_business()
+        business = self.require_business_permission("finance", "create")
         serializer = AccountSerializer(data=request.data)
         serializer.is_valid(raise_exception=True)
         serializer.save(business=business)
@@ -41,7 +36,7 @@ class AccountListView(BusinessOwnedMixin, APIView):
 
 class AccountDetailView(BusinessOwnedMixin, APIView):
     def get_object(self, business_id, pk):
-        self.get_business()
+        self.require_business_permission("finance", "view")
         return get_object_or_404(Account, business_id=business_id, pk=pk)
 
     def get(self, request, business_id, id):
@@ -49,6 +44,7 @@ class AccountDetailView(BusinessOwnedMixin, APIView):
         return Response(AccountSerializer(obj).data)
 
     def put(self, request, business_id, id):
+        self.require_business_permission("finance", "update")
         obj = self.get_object(business_id, id)
         serializer = AccountSerializer(obj, data=request.data)
         serializer.is_valid(raise_exception=True)
@@ -56,6 +52,7 @@ class AccountDetailView(BusinessOwnedMixin, APIView):
         return Response(serializer.data)
 
     def patch(self, request, business_id, id):
+        self.require_business_permission("finance", "update")
         obj = self.get_object(business_id, id)
         serializer = AccountSerializer(obj, data=request.data, partial=True)
         serializer.is_valid(raise_exception=True)
@@ -63,6 +60,7 @@ class AccountDetailView(BusinessOwnedMixin, APIView):
         return Response(serializer.data)
 
     def delete(self, request, business_id, id):
+        self.require_business_permission("finance", "delete")
         obj = self.get_object(business_id, id)
         obj.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
@@ -70,12 +68,12 @@ class AccountDetailView(BusinessOwnedMixin, APIView):
 
 class JournalListView(BusinessOwnedMixin, APIView):
     def get(self, request, business_id):
-        self.get_business()
+        self.require_business_permission("finance", "view")
         qs = Journal.objects.filter(business_id=business_id)
         return Response(JournalSerializer(qs, many=True).data)
 
     def post(self, request, business_id):
-        business = self.get_business()
+        business = self.require_business_permission("finance", "create")
         serializer = JournalSerializer(data=request.data)
         serializer.is_valid(raise_exception=True)
         serializer.save(business=business)
@@ -84,7 +82,7 @@ class JournalListView(BusinessOwnedMixin, APIView):
 
 class JournalDetailView(BusinessOwnedMixin, APIView):
     def get_object(self, business_id, pk):
-        self.get_business()
+        self.require_business_permission("finance", "view")
         return get_object_or_404(Journal, business_id=business_id, pk=pk)
 
     def get(self, request, business_id, id):
@@ -92,6 +90,7 @@ class JournalDetailView(BusinessOwnedMixin, APIView):
         return Response(JournalSerializer(obj).data)
 
     def put(self, request, business_id, id):
+        self.require_business_permission("finance", "update")
         obj = self.get_object(business_id, id)
         if obj.status != Journal.Status.DRAFT:
             return Response(
@@ -104,6 +103,7 @@ class JournalDetailView(BusinessOwnedMixin, APIView):
         return Response(serializer.data)
 
     def patch(self, request, business_id, id):
+        self.require_business_permission("finance", "update")
         obj = self.get_object(business_id, id)
         if obj.status != Journal.Status.DRAFT:
             return Response(
@@ -116,6 +116,7 @@ class JournalDetailView(BusinessOwnedMixin, APIView):
         return Response(serializer.data)
 
     def delete(self, request, business_id, id):
+        self.require_business_permission("finance", "delete")
         obj = self.get_object(business_id, id)
         if obj.status != Journal.Status.DRAFT:
             return Response(
@@ -128,6 +129,7 @@ class JournalDetailView(BusinessOwnedMixin, APIView):
 
 class JournalPostView(BusinessOwnedMixin, APIView):
     def post(self, request, business_id, id):
+        self.require_business_permission("finance", "update")
         obj = self.get_object(business_id, id)
         if obj.status != Journal.Status.DRAFT:
             return Response(
@@ -139,12 +141,13 @@ class JournalPostView(BusinessOwnedMixin, APIView):
         return Response(JournalSerializer(obj).data, status=status.HTTP_200_OK)
 
     def get_object(self, business_id, pk):
-        self.get_business()
+        self.require_business_permission("finance", "view")
         return get_object_or_404(Journal, business_id=business_id, pk=pk)
 
 
 class JournalReverseView(BusinessOwnedMixin, APIView):
     def post(self, request, business_id, id):
+        self.require_business_permission("finance", "update")
         obj = self.get_object(business_id, id)
         if obj.status != Journal.Status.POSTED:
             return Response(
@@ -156,20 +159,20 @@ class JournalReverseView(BusinessOwnedMixin, APIView):
         return Response(JournalSerializer(obj).data, status=status.HTTP_200_OK)
 
     def get_object(self, business_id, pk):
-        self.get_business()
+        self.require_business_permission("finance", "view")
         return get_object_or_404(Journal, business_id=business_id, pk=pk)
 
 
 class JournalEntryListView(BusinessOwnedMixin, APIView):
     def get(self, request, business_id, journal_id):
-        self.get_business()
+        self.require_business_permission("finance", "view")
         qs = JournalEntry.objects.filter(
             journal__business_id=business_id, journal_id=journal_id
         )
         return Response(JournalEntrySerializer(qs, many=True).data)
 
     def post(self, request, business_id, journal_id):
-        business = self.get_business()
+        business = self.require_business_permission("finance", "create")
         journal = get_object_or_404(
             Journal, business_id=business_id, pk=journal_id
         )
@@ -191,26 +194,26 @@ class JournalEntryListView(BusinessOwnedMixin, APIView):
 
 class LedgerListView(BusinessOwnedMixin, APIView):
     def get(self, request, business_id):
-        self.get_business()
+        self.require_business_permission("finance", "view")
         qs = Ledger.objects.filter(business_id=business_id)
         return Response(LedgerSerializer(qs, many=True).data)
 
 
 class LedgerDetailView(BusinessOwnedMixin, APIView):
     def get(self, request, business_id, id):
-        self.get_business()
+        self.require_business_permission("finance", "view")
         obj = get_object_or_404(Ledger, business_id=business_id, pk=id)
         return Response(LedgerSerializer(obj).data)
 
 
 class ExpenseListView(BusinessOwnedMixin, APIView):
     def get(self, request, business_id):
-        self.get_business()
+        self.require_business_permission("finance", "view")
         qs = Expense.objects.filter(business_id=business_id)
         return Response(ExpenseSerializer(qs, many=True).data)
 
     def post(self, request, business_id):
-        business = self.get_business()
+        business = self.require_business_permission("finance", "create")
         serializer = ExpenseSerializer(data=request.data)
         serializer.is_valid(raise_exception=True)
         account = serializer.validated_data.get("account")
